fastapi/requirements.txt
fastapi
uvicorn  # ASGI server for fastapi
# psycopg2  # psql api
psycopg2-binary
